# agent/api.yaml
openapi: "3.0.0"
info:
  version: "0.1.0"
  title: "Squill Agent REST API"
  description: The REST API used by the Squill Client to interact with the Agent.
paths:
...
  /connections/defaults:
    get:
      summary: Get a connection defaults values.
      security:
        - ApiKeyAuth: []
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Connection"

        "401":
          description: Unauthorized
...
  /connections/{id}:
    get:
      summary: Get a connection from its identifier.
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Connection"
        "401":
          description: Unauthorized

        "404":
          description: Connection not found
    delete:
      summary: Delete a connection from its identifier.
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Successful operation
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Connection not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /connections/{id}/queries/run:
    post:
      summary: Run one or more queries on the connection.
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          description: The unique identifier of the connection.
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RunRequest"
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/QueryExecution"
        "401":
          description: Unauthorized
  /connections/{id}/datasources:
    get:
      summary: List the datasources available for the connection.
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Datasource"
        "401":
          description: Unauthorized
  /connections/{id}/history:
    get:
      summary: List the queries in the history of the connection.
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: next_page
          description: |
            The pagination information retrieved from the previous call.
            This is used to retrieve the next page of the history if the number of queries is too large to be fulfilled
            by the previous calls.
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueryHistoryPage"
        "401":
          description: Unauthorized
  /connections/{id}/queries/{query_id}:
    get:
      summary: Get a query.
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: query_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueryExecution"
        "401":
          description: Unauthorized
        "404":
          description: Query not found

    delete:
      summary: Delete a query.
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: query_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: string
        "401":
          description: Unauthorized
        "404":
          description: Query not found
  /connections/{id}/queries/{query_id}/data:
    get:
      summary: Get the data of a query execution.
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: query_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: string
        "401":
          description: Unauthorized
        "404":
          description: Query not found
  /connections/validate:
    post:
      summary: |
        Check the validity of a connection definition.

        The connection definition is validate by the agent to ensure that the connection can be established.
        If the connection can be established the agent will return the connection information including the default
        datasource and all other datasources available for the connection.
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Connection"
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Connection"
        "401":
          description: Unauthorized
...
components:
  schemas:
...
    RunRequest:
      description: The payload of the POST /connections/{id}/run endpoint.
      type: object
      required:
        - origin
        - datasource
        - buffer
      properties:
        origin:
          type: string
          description: The origin of the request ('terminal', 'worksheet', etc).
        datasource:
          type: string
          description: The name of the datasource to be used within the connection.
        buffer:
          type: string
          description: A text buffer containing 0 or more queries.
...
    Datasource:
      description: The description of a datasource (database for SQL engines).
      type: object
      required:
        - name
      properties:
        name:
          type: string
          description: The name of the datasource (e.g. "my_database").
        description:
          type: string
          description: The description of the datasource (e.g. "My database").
        size_in_bytes:
          type: integer
          format: int64
          description: The size of the datasource in bytes.
        hidden:
          type: boolean
          description: A flag indicating if the datasource is hidden in the user interface.
    Connection:
      description: Definition of a connection to datasources.
      type: object
      x-immerable: true
      required:
        - id
        - parent_id
        - owner_user_id
        - mode
        - name
        - save_password
      properties:
        id:
          type: string
          format: uuid
          description: The unique identifier of the connection.
        parent_id:
          type: string
          format: uuid
          description: The unique identifier of the parent resource in the catalog.

        owner_user_id:
          type: string
          format: uuid
          description: The unique identifier of the user that owns the connection.

        driver:
          type: string
          description: The name of the driver used to connect to the datasource.

        name:
          type: string
          description: The name of the connection.

        alias:
          type: string
          format: identifier
          description: The alias of the connection.

        description:
          type: string
          description: The description of the connection.

        mode:
          $ref: "#/components/schemas/ConnectionMode"

        host:
          type: string
          description: The host of the connection.if the connection mode is "host".

        port:
          type: integer
          format: int16
          minimum: 0
          description: The port of the connection.if the connection mode is "host".

        socket:
          type: string
          description: The socket of the connection.if the connection mode is "socket".

        file:
          type: string
          description: The file of the connection.if the connection mode is "file".

        connection_string:
          type: string
          description: The connection string of the connection.if the connection mode is "connection_string".

        uri:
          type: string
          description: The URI of the connection.if the connection mode is "uri".

        username:
          type: string
          description: The username for authentication.

        password:
          type: string
          description: The password for authentication.

        save_password:
          type: boolean
          description: Whether the password should be saved.

        default_datasource:
          type: string
          description: The name of the default datasource to use.

        datasources:
          type: array
          description: The list of datasources available for the connection.
          items:
            $ref: "#/components/schemas/Datasource"

        options:
          type: object
          description: Additional options of the connection.
          additionalProperties:
            type: string
    ConnectionMode:
      description: The mode of a connection.
      type: string
      oneOf:
        - title: host
          description: The connection is a host-based connection.
        - title: socket
          description: The connection is a socket-based connection.
        - title: file
          description: The connection is a file-based connection.
        - title: connection_string
          description: The connection is a connection string-based connection.
        - title: uri
          description: The connection is a URI-based connection.
...
    Error:
      type: object
      required:
        - status
        - code
        - message
      properties:
        status:
          description: The HTTP status code
          type: integer
        code:
          description: A machine readable error code such as `invalid_parameter`.
          type: string
        message:
          description: A human readable error message
          type: string
...
    QueryExecutionStatus:
      description: |
        The status of a query execution.

        The following transitions are allowed:
        - `pending` -> `running` | `cancelled` | `deleted`
        - `running` -> `completed` | `failed` | `cancel_requested` | `delete_requested`
        - `completed` -> `deleted`
        - `failed` -> `deleted`
        - `cancel_requested` -> `cancelled` | `delete_requested`        
        - `cancelled` -> `deleted`
        - `delete_requested` -> `deleted`
        - `deleted` -> final state, no further transitions are allowed.

        The statuses `cancel_requested` and `delete_requested` are used to indicate that the query execution is still
        running but the user requested to cancel or delete the query execution. The query execution will transition to
        `cancelled` or `deleted` once the query execution is stopped.
      type: string
      oneOf:
        - title: pending
          description: Waiting to be executed.
        - title: running
          description: Execution in progress.
        - title: completed
          description: Execution completed successfully.
        - title: failed
          description: Execution failed.
        - title: cancel_requested
          description: |
            Cancellation requested.
            The cancellation of the query has been requested but the query is still running.
            Once the execution of the query is stopped the status will be updated to `cancelled`.
        - title: cancelled
          description: Execution was cancelled.
        - title: delete_requested
          description: |
            Deletion requested.
            The deletion of the query has been requested but the query is still running.
            Once the execution of the query is stopped the status will be updated to `deleted`.
        - title: deleted
          description: |
            Query deleted.
            This status is used to indicate that the query execution was removed from the history and the data
            associated to it on the file system can be deleted. Once the data deleted the query execution deletion is
            considered completed and the record in the query history can be removed.
    QueryExecutionError:
      description: An error message from a query execution.
      type: object
      properties:
        message:
          type: string
        line:
          type: integer
          format: int32
          minimum: 0
        column:
          type: integer
          format: int32
          minimum: 0
...
    QueryExecution:
      description: The execution of a query.
      type: object
      required:
        - id
        - revision
        - connection_id
        - user_id
        - text
        - hash
        - status
        - created_at
        - affected_rows
        - execution_time
        - origin
        - with_result_set
        - storage_bytes
        - storage_rows
      properties:
        id:
          description: The unique identifier of the query execution.
          type: string
          format: uuid
        revision:
          description: |
            The revision number of the query execution.

            The revision number is used to track the changes to the query execution. It is incremented each time the
            query execution is updated. Because the client receive updates of the query execution via different channels
            (HTTP and WebSocket) there is no guarantee that the last update received is the most recent. By using the
            revision number the client can avoid overwriting a more recent update with an older one.
            At creation the revision number is 0.
          type: integer
          format: int32
          minimum: 0
        connection_id:
          description: The unique identifier of the connection used to execute the query.
          type: string
          format: uuid
        datasource:
          description: The name of the datasource within the connection used to execute the query.
          type: string
        user_id:
          description: The unique identifier of the user that executed the query.
          type: string
          format: uuid
        username:
          description: |
            The name of the user that executed the query.
            This value is not stored in the database but is added to the query execution when it is returned to the
            client.
          type: string
        text:
          description: The text of the query.
          type: string
        hash:
          description: |
            The hash of the text of the query after normalization.

            The hash is used to determine if two queries are the same regardless of their formatting. The hash is 
            computed by normalizing the query into tokens.
          type: integer
          format: int64
          minimum: 0
        with_result_set:
          description: |
            A flag indicating if the query is a result set returning query.

            This flag is used to determine if the query execution may return the result set or not.

            Examples of result set returning queries are:
            - `SELECT``: The primary statement that retrieves rows from one or more tables.
            - `SHOW``: A statement that shows information about databases, tables, or other objects.
            - `INSERT ... RETURNING`: In some databases (like PostgreSQL), `INSERT``, `UPDATE``, and `DELETE`` can 
               return rows when combined with the `RETURNING` clause.
          type: boolean
        status:
          description: The status of the query execution.
          $ref: "#/components/schemas/QueryExecutionStatus"
        error:
          description: The error message if the query execution failed.
          $ref: "#/components/schemas/QueryExecutionError"
        executed_at:
          description: |
            The date and time when the query was executed.

            This is the time the query was submitted to the agent but the time when the query was submitted to the
            driver
          type: string
          format: date-time
        created_at:
          description: |
            The date and time when the query execution was created.

            This is the time when the query was submitted to the agent.
          type: string
          format: date-time
        affected_rows:
          description: The number of rows affected by the query.
          type: integer
          format: int64
          minimum: 0
        execution_time:
          description: |
            The time it took to execute the query in seconds.

            The time is captured in nanoseconds and converted to seconds using a 64-bit floating-point allowing for
            high precision on fast queries without loosing the ability to represent long running queries in seconds.
            This decision was made to keep that field usable in Javascript where the number type is a 64-bit 
            floating-point but can only represent integers up to 2^53 - 1 which would be only 2.5 hours in nanoseconds 
            before starting to loose precision. In addition seconds are more user friendly than nanoseconds.
          type: number
          format: double
          minimum: 0
        origin:
          description: |
            The origin of the query execution.

            The query can be originated from different origins like a terminal or a worksheet. In order to track the
            history of each of origin independently, the origin is stored in the query execution.

            Examples of origins are: `terminal`, `worksheet`, `e7ee76db-8758-4da4-bbce-242c8d1f3d63`, etc.
          type: string
        storage_bytes:
          description: The size of the result set on disk in bytes.
          type: integer
          format: int64
          minimum: 0
        storage_rows:
          description: The number of rows stored on disk.
          type: integer
          format: int64
          minimum: 0
        metadata:
          description: |
            A collection of key-value pairs that provide additional information about the query execution.

            - `schema`:
              The schema of the result set for queries with `with_result_set` set to `true`.
              The schema is a JSON representation of an Arrow schema using 
              [Apache Arrow JSON test data format](https://github.com/apache/arrow/blob/master/docs/source/format/Integration.rst#json-test-data-format)
              Having `with_result_set` set to `true` set to true doesn't guarantee that the schema will be present, the 
              schema is only present if the query execution was successful.
          type: object
          additionalProperties:
            type: string
...
    QueryHistoryPage:
      description: The response of the GET /connections/{id}/history/list endpoint.
      type: object
      required:
        - queries
      properties:
        queries:
          description: The list of queries in the history.
          type: array
          items:
            $ref: "#/components/schemas/QueryExecution"
        next_page:
          description: The pagination information for the next page.
          type: string
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-Api-Key

    BearerAuth:
      type: http
      scheme: bearer
security:
  - ApiKeyAuth: []
  - BearerAuth: []